Move custom SQL functions into separate file
